backend/requirements.txt
django
django-jazzmin
python-dotenv
psycopg2-binary
djangorestframework-simplejwt
